#4 Implement: RestController Class Exception Handle, Seller 권한 등록시 SellerInfo 생성
1. IdNotFoundException - 존재하지 않는 id로 query할 경우 발생
2. AlreadyExistException - 이미 존재하는 리소스를 생성할 경우 발생
3. UserIdNotMatchedException - 상품 등록한 User와 수정/삭제하려는 User가 다를 경우 발생
4. Seller 권한 등록시 SellerInfo 생성안하던 부분 더미 데이터로 생성하게끔 수정

+++ src/main/java/kr/co/mash_up/nine_tique/service/CategorySservice.java
@@ -1,7 +1,11 @@
 package kr.co.mash_up.nine_tique.service;
 
 import kr.co.mash_up.nine_tique.domain.Category;
+import kr.co.mash_up.nine_tique.exception.AlreadyExistException;
+import kr.co.mash_up.nine_tique.exception.IdNotFoundException;
 import kr.co.mash_up.nine_tique.repository.CategoryRepository;
+import kr.co.mash_up.nine_tique.vo.CategoryRequestVO;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,6 +16,7 @@ import java.util.List;
  * Category와 관련된 비즈니스 로직 처리
  */
 @Service
+@Slf4j
 public class CategorySservice {
 
     @Autowired
@@ -23,28 +28,44 @@ public class CategorySservice {
     }
 
     @Transactional(readOnly = true)
-    public Category findOne(Long id) {
-        return categoryRepository.findOne(id);
+    public Category findById(Long id) {
+        Category category = categoryRepository.findOne(id);
+
+        if (category == null) {
+            throw new IdNotFoundException("category find by id -> category not found");
+        }
+        return category;
     }
 
     @Transactional
     public Category update(Long id, Category newCategory) {
         Category oldCateogry = categoryRepository.findOne(id);
-        if(oldCateogry != null){
-            oldCateogry.setMain(newCategory.getMain());
-            oldCateogry.setSub(newCategory.getSub());
+
+        if (oldCateogry == null) {
+            throw new IdNotFoundException("category update -> category not found");
         }
 
+        oldCateogry.setMain(newCategory.getMain());
+        oldCateogry.setSub(newCategory.getSub());
         return categoryRepository.save(oldCateogry);
     }
 
     @Transactional
-    public Category create(Category category) {
-        return categoryRepository.save(category);
+    public Category create(CategoryRequestVO requestVO) {
+        Category oldCategory = categoryRepository.findByMainAndSubAllIgnoreCase(requestVO.getMain(), requestVO.getSub());
+
+        if (oldCategory != null) {  // 이미 존재할 경우
+            throw new AlreadyExistException("category create -> category already exist");
+        }
+        return categoryRepository.save(requestVO.toCategoryEntity());
     }
 
     @Transactional
     public void delete(Long id) {
+        Category oldCategory = categoryRepository.findOne(id);
+        if (oldCategory == null) {
+            throw new IdNotFoundException("category delete -> category not found");
+        }
         categoryRepository.delete(id);
     }
